Rename "inline" refactoring to align with "extract"
